+++ src/rgw/rgw_multi.cc
@@ -15,7 +15,7 @@
 #include "rgw_sal_rados.h"
 
 #include "services/svc_sys_obj.h"
-#include "services/svc_tier_rados.h"
+#include "services/rados/svc_tier_rados.h"
 
 #define dout_subsys ceph_subsys_rgw
