+++ src/rgw/rgw_putobj_processor.h
@@ -19,7 +19,7 @@
 
 #include "rgw_putobj.h"
 #include "services/rados/svc_rados.h"
-#include "services/svc_tier_rados.h"
+#include "services/rados/svc_tier_rados.h"
 #include "rgw_sal.h"
 #include "rgw_obj_manifest.h"
