+++ src/rgw/rgw_reshard.cc
@@ -18,7 +18,7 @@
 
 #include "services/svc_zone.h"
 #include "services/svc_sys_obj.h"
-#include "services/svc_tier_rados.h"
+#include "services/rados/svc_tier_rados.h"
 
 #define dout_context g_ceph_context
 #define dout_subsys ceph_subsys_rgw
