+++ src/rgw/services/rados/svc_tier_rados.cc
