+++ src/test/rgw/test_rgw_obj.cc
@@ -16,7 +16,7 @@
 #include "common/Formatter.h"
 #include "rgw/rgw_common.h"
 #include "rgw/rgw_rados.h"
-#include "rgw/services/svc_tier_rados.h"
+#include "rgw/services/rados/svc_tier_rados.h"
 #include "test_rgw_common.h"
 #include <gtest/gtest.h>
